backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-jose
passlib[bcrypt]
redis
requests
#tensorflow
python-dotenv
pydantic
langchain
langgraph
langchain-openai
apscheduler
twilio